Correctly store targetNodeId and fees when using Trampoline
This is ugly but works (except in one corner-case).
This is a temporary measure: the right way to fix this is to update the DB schema
to correctly account for Trampoline data.

Since DB schema changes are painful, we will aggregate all of them in one change later.
When we do that we will make this more elegant.