Rework features
Add types to allow cleaner dependency validation.
Most of the time we don't care whether a feature is activated as
optional or mandatory, which caused duplicate code.
This is now handled more cleanly.

It also paves the way to annotate features with the places they
should be advertised (Init vs NodeAnn vs ChannelAnn vs invoice).