Keep connection to peer even with no channels
Phoenix requires a connection to a remote peer for several features
to function properly (pay-to-open, swap-in, ...). When initializing
a new wallet, we have a connection but no channels, and the current
logic will prevent the node to reconnect to its peer in case the
connection is lost.

This commit adds special cases for the Phoenix app.