rework `AvailableBalanceChanged` events
Instead of emitting this event when we send a signature, we emit it when
our `availableBalanceForSend` actually changes. This happens:
- when we send a new `update_add` or `update_fee`;
- when we receive a `commit_sig`, which may acknowledge one or several
`update_fulfill`, `update_fail_*`, or `update_fee` sent by our peer.

We choose to only emit this event in `NORMAL` state, because its goal is
to give information about what payments we can make, which can only
happen in that state.

NB: other events `ChannelSignatureSent` and `ChannelSignatureReceived` give
a different type of information, and are sent in all states where
signatures are exchanges, not only in `NORMAL`.

The field `localBalance` has been moved to a function and is ambiguous
(there are multiple ways of computing our balance. We should probably
consider removing it altogether, it is only used for audit purposes and
arguably not very useful there.