Electrum: allow watcher to watch for mempool transactions (#1298)
* Electrum: allow watcher to watch for mempool transactions

Watcher now handles WatchConfirmed watches where min depth 
is set to 0: the watch event will be sent when the tx enters the 
mempool of the bitcoin node our Electrum server is connected to.

For 0-conf channel, use scids with a height of 0 and a tx index 
generated from the first 16 bytes of the funding txid. This gives us 
unique ids that can still be identified as 0-conf channel.