Funder reserve for future fee increase
See https://github.com/lightningnetwork/lightning-rfc/issues/728

Add an additional reserve on the funder to prevent emptying and then
being stuck with an unusable channel.

As fundee we don't verify funders comply with that change.
We may enforce it in the future when we're confident the network as a
whole enforces that.