remote: front now handles ping/sync
Peer has been split in two and now handles only channel related stuff.

A new PeerConnection class is in charge of managing the BOLT 1 part
(init handshake, pings) and has the same lifetime as the underlying
connection.

Two possible next steps:
- have `TransportHandler` be a child of `PeerConnection`; this is easy
for outgoing connection, but not straightforward for incoming
connections because we don't know what the remote node id is before the
crypto hasndshake. We could work around that by either (1) making the
`remoteNodeId` an attribute of the state of `PeerConnection` instead of
a constructor argument or (2) read the first few bytes of the incoming
connection before the crypto handshake to know the node id before it is
authenticated.
- put the reconnection logic outside of the `Peer` to simplify it even
further.