Add missing cases to PostRestart
This commit handles two important cases post-restart.

The first one is that we should ignore outgoing HTLCs that have already
been settled upstream (which can happen when downstream is closing).

The second one is that we should handle AddHtlcFailed messages which
will be sent when downstream is closing and we realize an HTLC will either
never reach the blockchain, or has been correctly timed out on-chain.