Add htlc details in AddHtlcFailed
In some cases, we use AddHtlcFailed for HTLCs that have been added
to the commit tx but resulted in an error.

This can happen when an error occurs during the revocation/signing
process, or when the HTLC is timed out on-chain.

It's useful in such cases to provide the HTLC in the event.