Clean up usage of AddHtlcFailed
We were abusing AddHtlcFailed in some cases where an outgoing HTLC
was correctly added, but was later failed on-chain (timed out or overridden).

AddHtlcFailed is now only one case of a failure that we need to
propagate upstream. HTLC timeout and override are other cases of
the more general concept of local htlc failure (for which we didn't receive
an UpdateFailHtlc from our remote but still need to behave similarly).