separated connection logic from Peer
The `ReconnectionTask` actor handles outgoing connections to a peer. The goal is to free
the `Peer` actor from the reconnection logic and have it just react to already established
connections, independently of whether those connections are incoming or outgoing.

The base assumption is that the `Peer` will send its state transitions
to the `ReconnectionTask` actor.

This is more complicated than it seems and there are various corner
cases to consider:
- multiple available addresses
- concurrent outgoing connections and conflict between
automated/user-requested attempts
- concurrent incoming/outgoing connections and risk of reconnection
loops
- etc.