split the `Peer` in two
* front now handles ping/sync

Peer has been split in two and now handles only channel related stuff.

A new PeerConnection class is in charge of managing the BOLT 1 part
(init handshake, pings) and has the same lifetime as the underlying
connection.

Two possible next steps:
- have `TransportHandler` be a child of `PeerConnection`; this is easy
for outgoing connection, but not straightforward for incoming
connections because we don't know what the remote node id is before the
crypto hasndshake. We could work around that by either (1) making the
`remoteNodeId` an attribute of the state of `PeerConnection` instead of
a constructor argument or (2) read the first few bytes of the incoming
connection before the crypto handshake to know the node id before it is
authenticated.
- put the reconnection logic outside of the `Peer` to simplify it even
further.

* remove `Authenticator`

Instead of first authenticating a connection, then passing it to the
`PeerConnection` actor, we pass the connection directly to the
`PeerConnection` and let it handle the crypto handshake, before the LN
init. This removes a central point of management and makes things easier
to reason about. As a side effect, the `TransportHandler` actor is now a
child of `PeerConnection` which gives us a guarantee that it dies when
its parent dies.

* separated connection logic from `Peer`

The `ReconnectionTask` actor handles outgoing connections to a peer. The
goal is to free
the `Peer` actor from the reconnection logic and have it just react to
already established
connections, independently of whether those connections are incoming or
outgoing.

The base assumption is that the `Peer` will send its state transitions
to the `ReconnectionTask` actor.

This is more complicated than it seems and there are various corner
cases to consider:
- multiple available addresses
- concurrent outgoing connections and conflict between
automated/user-requested attempts
- concurrent incoming/outgoing connections and risk of reconnection
loops
- etc.

Co-Authored-By: Bastien Teinturier <31281497+t-bast@users.noreply.github.com>