Revert Electrum Tor disable SSL
Revert #1278. The correct way to handle that is to add more ways for
users to choose their SSL behavior in phoenix/eclair-mobile.

In the long run it will be via more checkboxes / a better configuration
panel, but at first it can simply be moving #1278's logic directly in both
Phoenix and eclair-mobile (which I'll do in follow-up PRs).