Router channel balance (#1382)
Router uses channel events (LocalChannelUpdate and AvailableBalanceChanged)
to track the balance of local channels.

This information will be used in path-finding to improve path-finding,
especially in the MPP case.

The information is added to the graph structures, but it's not used yet in
path-finding. Some A/B testing will be needed before we can use those
safely for the path-finding algorithm.