make phoenix resist to multi device
There are several scenarii when phoenix starts up:
(a) peer sends a `channel_reestablish` for an unknown channel
    This is the restore use case. Phoenix will decrypt the attached
    `channel_data` and restore the channel
(b) peer sends a `channel_reestablish` for a known channel
    Phoenix decrypts the attached `channel_data` and compares them to
    the data it gots. Then:
    (1) if the remote data is more recent than local ones, Phoenix will
    use the remote data. This is the multi device use case.
    (2) otherwise, Phoenix keeps its data. This is the normal use case.
(c) peer doesn't send a `channel_reestablish` for a known channel.
    This would happen if the peer lost its data. In that case Phoenix
    will wait forever for a `channel_reestablish` and the channel will
    stay in OFFLINE. User can decide to force close the channel.

Until now, we were only handling (a) and (b)(2). (b)(1) would lead
to a force close of the channel, because Phoenix would enter the data
recovery scenario.

We now make Phoenix *react* to the peer's `channel_reestablish`, which
creates the (c) corner case, which seems acceptable. If the peer lost
its data, the only way out is to force close the channel anyway.