Scale fulfill safety window based on amount at risk
When an upstream peer is not responsive to htlc fulfills, there is a risk
that they will try to cheat us by force-closing the channel and redeeming
the htlc-timeout on-chain.

To avoid that, we previously used a constant security window of 6 blocks
to force-close the channel ourselves (which gives us time to redeem the
htlc-success on-chain).

Instead of using a constant window, it makes sense to scale it based on
the amount at risk. Coupled with the `max-htlc-value-in-flight-msat`
restriction, it provides a good adaptive defence against those attacks.