remove the `override-features` conf parameter
The per-node override was only used in the `init` message, and not in
e.g. channel operations, which means it was completely broken.

The use-case for this is Eclair Mobile, because we only sync the routing
table from the ACINQ node (because there is no public channel
validation as it is too expensive on mobile devices).

I'm not sure what to do with this. Should we only have this feature on
the `android` branch?

Related to #1361.