Fix flaky integration test
The downstream HTLC-timeout integration test was sometimes hanging waiting
for the local commit tx to appear inside the mempool.

The reason was that the remote peer was also trying to get its version of
the commit tx in the mempool, and when it won that race we weren't testing
the right thing.

Simply disconnecting the two nodes fixes the issue as it ensure only the
local node will be broadcasting his commit tx.